Move raise_if_invalid_coordinates into utils.py
